requirements.txt
numpy==1.22.3
opencv_python==4.5.5.64
pandas==1.4.2
Pillow==9.1.1
streamlit==1.9.0
tensorflow==2.9.0
